added view event listener to navigate to detail fragment passing along with ID

+++ app/src/main/java/com/example/yelpedapp/api/BusinessDTO.kt
@@ -7,18 +7,18 @@ import com.squareup.moshi.JsonClass
 data class BusinessDTO(
     @Json(name = "id") val id: String,
     @Json(name = "name") val name: String,
-    @Json(name = "rating") val rating: Double,
+    @Json(name = "rating") val rating: Float,
     @Json(name = "price") val price: String?,
     @Json(name = "phone") val phone: String,
     @Json(name = "alias") val alias: String,
     @Json(name = "is_closed") val isClosed: Boolean,
     @Json(name = "categories") val categories: List<CategoryDTO>,
-    @Json(name = "review_count") val reviewCount: Long,
+    @Json(name = "review_count") val reviewCount: Int,
     @Json(name = "url") val url: String,
     @Json(name = "coordinates") val coordinates: CenterDTO,
     @Json(name = "image_url") val imageURL: String,
     @Json(name = "location") val location: LocationDTO,
-    @Json(name = "distance") val distance: Double,
+    @Json(name = "distance") val distance: Float,
     @Json(name = "transactions") val transactions: List<String>
 )
